+++ src/widgets/input/TbInputInline.php
@@ -88,7 +88,7 @@ class TbInputInline extends TbInputVertical
 		echo $this->form->typeAheadField($this->model, $this->attribute, $this->data, $this->htmlOptions);
 		echo $this->getAppend();
 	}
-    
+
      /**
      * Renders a datepicker field.
      * @return string the rendered content
@@ -110,6 +110,39 @@ class TbInputInline extends TbInputVertical
         echo $this->getPrepend();
         $this->widget(
             'bootstrap.widgets.TbDatePicker',
+            array(
+                'model' => $this->model,
+                'attribute' => $this->attribute,
+                'options' => isset($options) ? $options : array(),
+                'events' => isset($events) ? $events : array(),
+                'htmlOptions' => $this->htmlOptions,
+            )
+        );
+        echo $this->getAppend();
+        echo $this->getError() . $this->getHint();
+    }
+
+     /**
+     * Renders a datetimepicker field.
+     * @return string the rendered content
+     * @author Hrumpa
+     */
+    protected function datetimepickerField()
+    {
+        if (isset($this->htmlOptions['options'])) {
+            $options = $this->htmlOptions['options'];
+            unset($this->htmlOptions['options']);
+        }
+
+        if (isset($this->htmlOptions['events'])) {
+            $events = $this->htmlOptions['events'];
+            unset($this->htmlOptions['events']);
+        }
+
+        echo $this->setPlaceholder();
+        echo $this->getPrepend();
+        $this->widget(
+            'bootstrap.widgets.TbDateTimePicker',
             array(
                 'model' => $this->model,
                 'attribute' => $this->attribute,